+++ adapters/hot-refresh-adapter-lombok/src/main/java/com/hyf/hotrefresh/adapter/lombok/LombokAdapterInstaller.java
@@ -2,7 +2,7 @@ package com.hyf.hotrefresh.adapter.lombok;
 
 import com.hyf.hotrefresh.core.install.Installer;
 import com.hyf.hotrefresh.core.memory.AnnotationProcessorCompositeClassLoader;
-import com.hyf.hotrefresh.core.util.InfrastructureJarClassLoader;
+import com.hyf.hotrefresh.core.classloader.InfrastructureJarClassLoader;
 import com.hyf.hotrefresh.core.util.Util;
 
 import java.lang.reflect.Field;