+++ adapters/hot-refresh-adapter-mapstruct/src/main/java/com/hyf/hotrefresh/adapter/mapstruct/MapStructAdapterInstaller.java
@@ -1,7 +1,7 @@
 package com.hyf.hotrefresh.adapter.mapstruct;
 
 import com.hyf.hotrefresh.core.install.Installer;
-import com.hyf.hotrefresh.core.util.InfrastructureJarClassLoader;
+import com.hyf.hotrefresh.core.classloader.InfrastructureJarClassLoader;
 
 /**
  * @author baB_hyf