+++ adapters/hot-refresh-adapter-mapstruct/src/test/java/com/hyf/hotrefresh/adapter/mapstruct/MapStructAdapterInstallerTests.java
@@ -1,6 +1,6 @@
 package com.hyf.hotrefresh.adapter.mapstruct;
 
-import com.hyf.hotrefresh.core.util.InfrastructureJarClassLoader;
+import com.hyf.hotrefresh.core.classloader.InfrastructureJarClassLoader;
 import org.junit.Before;
 import org.junit.Test;